chore(drizzle): exclude neon_auth.* from drizzle-kit migration scope
Drizzle-kit's tablesFilter/schemaFilter options only filter DB
introspection (push/pull), not generate's schema-vs-snapshot diffing
— probe-tested both. The actual fix is structural:

  1. Move neonAuthUsers/neonAuthSchema to a separate file that
     drizzle.config.ts:schema does not point at (external-schema.ts).
  2. schema.ts imports neonAuthUsers but does not re-export. Drizzle-kit
     only enumerates exports of the configured schema file, so the
     foreign auth table becomes invisible to its diffing while
     references(() => neonAuthUsers.id, ...) still resolves at runtime
     for FK SQL emission.
  3. Hand-trim 0007_snapshot.json to remove the neon_auth.user table
     entry and empty the schemas object, so drizzle-kit doesn't see
     a "snapshot has it, schema doesn't" diff and try to emit
     DROP TABLE "neon_auth"."user" CASCADE; DROP SCHEMA "neon_auth";

drizzle.config.ts stays bare — the filter flags don't help.

Verified: clean-tree drizzle-kit generate produces "No schema changes";
probe (add column to neonAuthUsers in external-schema.ts) emits no
migration; sanity probe (add column to a public table) emits the
expected ALTER; pnpm check-types:web exits 0.

Closes docs/remaining_work.md §4.11.
