+++ apps/web/drizzle/meta/0007_snapshot.json
@@ -154,61 +154,6 @@
       "checkConstraints": {},
       "isRLSEnabled": false
     },
-    "neon_auth.user": {
-      "name": "user",
-      "schema": "neon_auth",
-      "columns": {
-        "id": {
-          "name": "id",
-          "type": "uuid",
-          "primaryKey": true,
-          "notNull": true
-        },
-        "name": {
-          "name": "name",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": false
-        },
-        "email": {
-          "name": "email",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": false
-        },
-        "email_verified": {
-          "name": "email_verified",
-          "type": "boolean",
-          "primaryKey": false,
-          "notNull": false
-        },
-        "image": {
-          "name": "image",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": false
-        },
-        "created_at": {
-          "name": "created_at",
-          "type": "timestamp",
-          "primaryKey": false,
-          "notNull": false
-        },
-        "updated_at": {
-          "name": "updated_at",
-          "type": "timestamp",
-          "primaryKey": false,
-          "notNull": false
-        }
-      },
-      "indexes": {},
-      "foreignKeys": {},
-      "compositePrimaryKeys": {},
-      "uniqueConstraints": {},
-      "policies": {},
-      "checkConstraints": {},
-      "isRLSEnabled": false
-    },
     "public.order_lines": {
       "name": "order_lines",
       "schema": "",
@@ -918,9 +863,7 @@
       ]
     }
   },
-  "schemas": {
-    "neon_auth": "neon_auth"
-  },
+  "schemas": {},
   "sequences": {},
   "roles": {},
   "policies": {},
